+++ lib/server/serverStorage.ts
@@ -11,10 +11,12 @@
  *
  */
 
+// @ts-ignore 
+// Do not use this type in code, so it won't be emitted in the declaration output
+import NodeCache from 'node-cache';
 import { SimpleStorage, StorageType, StorageUtil, Cookies } from '../types';
 import { AuthSdkError } from '../errors';
 // eslint-disable-next-line import/no-commonjs
-const NodeCache = require('node-cache'); // commonJS module cannot be imported without esModuleInterop
 
 // this is a SHARED memory storage to support a stateless http server
 const sharedStorage = typeof NodeCache === 'function' ? new NodeCache() : null;