+++ scripts/validate-bundles.sh
@@ -5,8 +5,18 @@ source ${OKTA_HOME}/${REPO}/scripts/setup.sh "v14.18.0"
 export TEST_SUITE_TYPE="junit"
 export TEST_RESULT_FILE_DIR="${REPO}/build2/reports/unit"
 
-if ! yarn test:bundle:esm; then
-  echo "validate esm bundle failed! Exiting..."
+if ! yarn test:bundle:esm:browser; then
+  echo "validate ESM browser bundle failed! Exiting..."
+  exit ${TEST_FAILURE}
+fi
+
+if ! yarn test:bundle:esm:node; then
+  echo "validate ESM node bundle failed! Exiting..."
+  exit ${TEST_FAILURE}
+fi
+
+if ! yarn test:bundle:cjs; then
+  echo "validate cjs bundle failed! Exiting..."
   exit ${TEST_FAILURE}
 fi