+++ test/apps/node-esm/index.js
@@ -1,0 +1,7 @@
+/* eslint-disable */
+import { OktaAuth } from "@okta/okta-auth-js";
+
+const oktaAuth = new OktaAuth({ 
+  issuer: 'https://xxx.okta.com',
+  clientId: '0oal89rzfrHjIVqQw5d6'
+});