+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - [#1113](https://github.com/okta/okta-auth-js/pull/1113) Updates types for `SigninWithCredentialsOptions` and `SignInOptions` to support `SP Initiated Auth`
 - [#1125](https://github.com/okta/okta-auth-js/pull/1125) IDX - Supports auto select methodType (when only one selection is available) for `authenticator-verification-data` remediation
+- [#1114](https://github.com/okta/okta-auth-js/pull/1114) Exposes ESM node bundle
+
+### Fixes
+
+- [#1114](https://github.com/okta/okta-auth-js/pull/1114) Fixes ESM browser bundle issue by only using ESM `import` syntax
 
 ## 6.1.0